Final Fantasy on OFFICIAL odds — WotC's own Collecting-article rates
First Magic set with publisher-published rates: WotC's 'Collecting FINAL
FANTASY' article states the Play Booster slot numerically (default rare 80% /
mythic 10% / borderless rare 8% / borderless mythic 1%) and puts a Through
the Ages card in one third of packs, split uncommon 63.25% / rare 29.75% /
mythic 7%. All transcribed verbatim into fin.json v2 (confidence OFFICIAL,
article as sourceUrl):

- 79 borderless rares/mythics move from display-only into their own
  borderless_rare/borderless_mythic tiers at the stated slot shares
- the 64 FCA cards join fin at NON-foil play prices under three bonus tiers,
  so the stated rarity split is modelled instead of flattened (guaranteed
  slots with fractional expected counts; schema relaxed accordingly)
- the ~1% FF-artist series stays display-only (not worth a misclassification)
  and the foil slot stays unmodelled, so EV still understates slightly

FIN Play Box: EV $101 -> $332 vs $230 market = +44.4% — an 8th site
positive, and the first carried by publisher-printed rates. Methodology's
Magic caveat now names the official-article exception.

+++ apps/web/src/lib/pullrates/schema.ts
@@ -33,7 +33,7 @@ const slotSchema = z.object({
 const guaranteedSlotSchema = z.object({
   label: z.string().min(1),
   rarity: z.string().min(1),
-  countPerPack: z.number().int().positive(),
+  countPerPack: z.number().positive(),
 });
 
 const boxGuaranteeSchema = z.object({
